start terraform removal
