--- build.sh
@@ -1,112 +1,0 @@
-#!/bin/bash
-
-source configs/process-config.env
-
-export PROCESS=$1
-
-make-bcm () {
-    CORES=$( nproc )
-    PER=0.8
-    CALC=$( echo "$NPROC*$PER" | bc )
-    cores=$(printf '%.0f' $CALC)
-
-    find $DATA_DIR/point-clouds/${STATE}/${WORKUNIT} -maxdepth 1 -name "*.laz" | \
-        xargs -P ${cores} -t -I % \
-        make bcm pc=% 
-}
-
-make-dsm () {
-    CORES=$(nproc)
-    find $DATA_DIR/bcm/${STATE}/${WORKUNIT}  -maxdepth 1 -name "*.laz" | \
-        xargs -P ${CORES} -t -I % \
-        make dsm pc=%
-
-    make vrt 
-}
-
-make-tin () {
-    CORES=$(nproc)
-    find $DATA_DIR/bcm/${STATE}/${WORKUNIT}  -maxdepth 1 -name "*.laz" | \
-        xargs -P ${CORES} -t -I % \
-        make tin pc=%
-
-    make vrt 
-}
-
-make-solar () {
-    CORES=$(nproc)
-    find $DATA_DIR/dsm/${STATE}/${WORKUNIT} -maxdepth 1 -name "*.tif" | \
-        xargs -P ${CORES} -t -I % \
-        make solar-average tif=% 
-
-    make vrt  
-}
-
-make-hillshade () {
-    CORES=$(nproc)
-    find $DATA_DIR/${PROCESS}/${STATE}/${WORKUNIT} -maxdepth 1 -name "*.tif" | \
-        xargs -P ${CORES} -t -I % \
-        make hillshade tif=% 
-    make vrt 
-}
-
-make-cog () {
-    CORES=$(nproc)
-    if [ $HS == "true" ] && [ $PROCESS != "solar" ]; then
-        data_dir=$DATA_DIR/${PROCESS}/${STATE}/${WORKUNIT}/hillshade
-    else
-        data_dir=$DATA_DIR/${PROCESS}/${STATE}/${WORKUNIT}
-    fi
-    find $data_dir -maxdepth 1 -name "*.tif" | \
-        xargs -P ${CORES} -t -I % \
-        make reproject tif=%
-    make cog
-}
-
-set-data-dir () {
-    if [[ $TYPE == "test" ]]; then
-        export DATA_DIR="test-data"
-    else
-        export DATA_DIR="data"
-    fi
-}
-
-## Run Processes
-set-data-dir
-processName="make-${PROCESS}" 
-if [[ $LOCATION = "remote" ]]; then
-    git clone --branch $GIT https://${TOKEN}@github.com/xycarto/wesm-surfaces.git
-    cp -r .creds wesm-surfaces
-    cd wesm-surfaces
-    echo -e "PROCESS=$1\n" >> configs/process-config.env
-    echo -e "DATA_DIR=$DATA_DIR\n" >> configs/process-config.env
-    echo $COG
-    if [[ $COG == "true" ]]; then
-        make download-files
-        make-cog
-    elif [[ $COG == "false" ]]; then
-        make download-files 
-        if [[ $HS == "true" ]]; then
-            make-hillshade
-        else
-            $processName
-        fi
-    fi
-elif [[ $LOCATION = "local" ]]; then
-    if [[ $COG == "true" ]]; then
-        make download-files
-        make-cog
-    elif [[ $COG == "false" ]]; then
-        make download-files 
-        if [[ $HS == "true" ]]; then
-            make-hillshade
-        else
-            $processName
-        fi
-    fi
-fi
-
-
-
-exit
-
