+++ configs/process-config.env
@@ -1,10 +1,2 @@
 WORKUNIT=CA_NoCAL_Wildfires_B1_2018
 STATE=California
-EC2=c5.2xlarge
-VOLUME_SIZE=50
-TYPE=test
-TEST_NUM=10
-LOCATION=local
-GIT=main
-HS=false
-COG=false
