--- docker/Dockerfile.terraform
@@ -1,32 +1,0 @@
-FROM ubuntu:20.04
-
-RUN apt-get update && apt-get install --no-install-recommends -y gnupg software-properties-common wget
-
-RUN wget -O- https://apt.releases.hashicorp.com/gpg | \
-    gpg --dearmor | \
-    tee /usr/share/keyrings/hashicorp-archive-keyring.gpg
-
-RUN gpg --no-default-keyring \
-    --keyring /usr/share/keyrings/hashicorp-archive-keyring.gpg \
-    --fingerprint
-
-RUN echo "deb [signed-by=/usr/share/keyrings/hashicorp-archive-keyring.gpg] \
-    https://apt.releases.hashicorp.com $(lsb_release -cs) main" | \
-    tee /etc/apt/sources.list.d/hashicorp.list
-
-RUN apt update && apt-get install --no-install-recommends -y terraform
-
-RUN apt-get update && apt-get --no-install-recommends -y install python3-pip
-RUN pip3 install awscli
-
-RUN apt-get update && apt-get --no-install-recommends -y install openssh-server
-
-RUN useradd -rm -d /home/ubuntu -s /bin/bash -g root -G sudo -u 1000 test 
-
-RUN  echo 'test:test' | chpasswd
-
-RUN service ssh start
-
-EXPOSE 22
-
-CMD ["/usr/sbin/sshd","-D"]
