+++ src/bcm-no-buff.py
@@ -1,10 +1,7 @@
 #!/usr/bin/env python
 import os
 import geopandas as gp
-import subprocess as sub
-from pyproj import CRS
 import sys
-import shutil
 from py_utils import *
 from globals import *
 
@@ -16,35 +13,15 @@ def main():
     index_file = gp.read_file(INDEX_FILE)
     index_row = index_file[index_file['file_name'] == os.path.basename(IN_FILE)]
 
-    filter_laz(IN_FILE, index_row)
-    
+    print(index_row.usgs_loc.values[0])
 
-def filter_laz(in_file, index_row):
-    print("Filtering...")   
-    bcm_file =  f"{BCM_DIR}/{os.path.basename(in_file)}"
-    crs = f"EPSG:{index_row.native_horiz_crs.values[0]}"
-    sub.call(
-        f"pdal -v 0 --nostream pipeline '{PIPELINE_FILTER}' \
-            --readers.las.filename='{in_file}' \
-            --writers.las.filename='{bcm_file}' \
-            --writers.las.a_srs='{crs}'",
-        shell=True,
-    )
+    get_usgs_file(s3, index_row.usgs_loc.values[0], IN_FILE, USGS_BUCKET)
 
-def merge_pc(clipped_array):       
-    print("Merging...")
-    merged_pc = f"{BCM_DIR}/{os.path.basename(IN_FILE).split('.')[0]}.laz"
-    sub.call(
-        f"""pdal -v 0 merge {" ".join(clipped_array)} {merged_pc}""",
-        shell=True,
-    )
-    return merged_pc
+    filter_laz(IN_FILE, index_row)    
 
-       
 if __name__ == "__main__":
 
     IN_FILE = sys.argv[1]
-    CLIP_DIR = f"{DATA_DIR}/clips/{os.path.basename(IN_FILE).split('.')[0]}"
         
     for d in [DATA_DIR, PC_DIR, BCM_DIR,INDEX_DIR]:
         os.makedirs(d, exist_ok=True)
