+++ src/globals.py
@@ -1,16 +1,16 @@
 import os 
 
 ## Environment vars from: configs/process-config.env
-DATA_DIR = os.environ.get("DATA_DIR")
 WORKUNIT = os.environ.get("WORKUNIT") 
 STATE = os.environ.get("STATE")
-TYPE = os.environ.get("TYPE")
-PROCESS = os.environ.get("PROCESS")
-TEST_NUM = int(os.environ.get("TEST_NUM"))
-HS = os.environ.get("HS")
-COG = os.environ.get("COG")
+# TYPE = os.environ.get("TYPE")
+# PROCESS = os.environ.get("PROCESS")
+# TEST_NUM = int(os.environ.get("TEST_NUM"))
+# HS = os.environ.get("HS")
+# COG = os.environ.get("COG")
 
 ## CONSTANTS
+DATA_DIR = 'data'
 USGS_BUCKET = "usgs-lidar"
 WESM_SURFACE_BUCKET = "xyc-wesm-surfaces"
 WESM_GRID_BUCKET = "xyc-wesm-grids"
@@ -32,8 +32,8 @@ INDEX_FILE = f"{INDEX_DIR}/{WORKUNIT}_index_4269.gpkg"
 GRID_INDEX_FILE = f"{GRID_INDEX_DIR}/{WORKUNIT}_index_4269.gpkg"
 
 ## PDAL Pipelines
-PIPELINE_CROP_FILTER = 'src/pipeline-templates/buffer-clip-filter-template.json'
-PIPELINE_FILTER = 'src/pipeline-templates/buffer-clip-filter-template-filter-only.json'
-PIPELINE_DSM = 'src/pipeline-templates/dsm_template.json'
+PIPELINE_CROP_FILTER = 'src/pipeline-templates/bcm-template.json'
+PIPELINE_FILTER = 'src/pipeline-templates/bcm-template-filter-only.json'
+PIPELINE_DSM = 'src/pipeline-templates/dsm-template.json'
 PIPELINE_TIN = 'src/pipeline-templates/tin-template.json'
