+++ src/pipeline-templates/bcm-template-crop-only.json
