+++ src/pipeline-templates/bcm-template-filter-only.json
