+++ src/pipeline-templates/bcm-template.json
