+++ src/pipeline-templates/dsm-template.json
