+++ src/pipeline-templates/unused/dem_gdal_template.json
