+++ src/pipeline-templates/unused/dem_template.json
