+++ src/pipeline-templates/unused/hexbin_template.json
