--- terraform/outputs.tf
@@ -1,14 +1,0 @@
-output "instance_id" {
-  description = "ID of the EC2 instance"
-  value       = aws_instance.app_server.id
-}
-
-output "instance_public_ip" {
-  description = "Public IP address of the EC2 instance"
-  value       = aws_instance.app_server.public_ip
-}
-
-output "instance_name" {
-  description = "Name of Instance"
-  value       = aws_instance.app_server[*].tags["Name"]
-}
